+++ Events.php
@@ -2,24 +2,12 @@
 
 namespace humhub\modules\socialshare;
 
-use Yii;
-use yii\helpers\Url;
-use humhub\modules\user\helpers\AuthHelper;
+use humhub\modules\socialshare\widgets\ShareLink;
+use yii\base\WidgetEvent;
 
 class Events
 {
-
-    public static function onTopMenuInit($event)
-    {
-        $event->sender->addItem([
-            'label' => 'SocialShare',
-            'url' => Url::toRoute('/socialshare/main/index'),
-            'icon' => '<i class="fa fa-sun-o"></i>',
-            'isActive' => (Yii::$app->controller->module && Yii::$app->controller->module->id == 'socialshare'),
-        ]);
-    }
-    
-    public static function onWallEntryLinksInit($event)
+    public static function onWallEntryLinksAfterRun(WidgetEvent $event)
     {
         if (!Yii::$app->user->isGuest || Yii::$app->user->identity && AuthHelper::isGuestAccessEnabled())
         {